+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.0.3] - 30.01.20
+
 ### Added
 
 - Support for attributes in `Newline`.
@@ -35,5 +37,6 @@
 - `Superstring` basic swift package structure.
 
 [Unreleased]: https://github.com/manuelCarlos/Superstring/compare/v0.0.1...HEAD
+[0.0.3]: https://github.com/manuelCarlos/Superstring/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/manuelCarlos/Superstring/compare/0.0.1...0.0.2
 [0.0.1]: https://github.com/manuelCarlos/Superstring/releases/tag/0.0.1